Display schedules and seats for a movie.
Move seats into schedule so that the data can go straight from clicked schedule to the displayed seat.

Add schedule prop to seats component to be able to show the seats from that schedule.

Map bookings to the correct schedule in schedules so that the clicked schedule always shows the correct amount of seats left for that schedule.